pyproject.toml
[tool.poetry]
name = "coding-with-roby"
version = "0.1.0"
description = ""
authors = ["Jefest1 <anyanf101@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.5"}
pydantic-settings = "^2.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
